backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.1
numpy==1.26.0
openai==1.2.4
faiss-cpu==1.10.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0